requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
nest_asyncio==1.6.0
httpx==0.25.2
redis==6.0.0